highlighter: don't crash when ctree is refreshed
When hexrays pushes a 'refresh_pseudocode' event, it most likely means
the ctree has been rebuilt with new objects. PseudocodeHighlighter
used to keep track of the currently highlighted block by keeping a
pointer the corresponding citem. Trying to use this now-discarded
citem during a refresh event was a recipe for all kinds of bad things.

PseudocodeHighlighter now keeps track of the current item with a pair
of the item's op and ea; when the tree gets refreshed it tries to
find the corresponding item. Much safer this way.